backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.10
requests==2.32.3
gunicorn==23.0.0
python-dotenv==1.1.0
blinker==1.9.0
apscheduler==3.10.4
